раздел 5 задания 1 и 2
