+++ js/functions.js
@@ -1,7 +1,7 @@
 const checksLength = (str, quantityChars) => str.length <= quantityChars;
 
 const checksPalindrome = (str) => {
-  const normalStr = (str.toLowerCase()).replaceAll(' ', '');
+  const normalStr = str.toLowerCase().replaceAll(' ', '');
   let reverseStr = '';
   for (let i = normalStr.length - 1; i >= 0; i--) {
     reverseStr += normalStr[i];
@@ -20,6 +20,25 @@ const extractsDigits = (str) => {
       strFromDigits += str[i];
     }
   }
-  return strFromDigits === '' ? NaN : strFromDigits;
+  return strFromDigits === '' ? NaN : +strFromDigits;
 };
+
+const checksMeeting = (startDay, endDay, startMeeting, durationMeeting) => {
+  const startDayArr = startDay.split(':');
+  const endDayArr = endDay.split(':');
+  const startMeetingArr = startMeeting.split(':');
+
+  const startDayValue = startDayArr.reduce((sum, element) => sum * 60 + (+element), 0);
+  const endDayValue = endDayArr.reduce((sum, element) => sum * 60 + (+element), 0);
+  const startMeetingValue = startMeetingArr.reduce((sum, element) => sum * 60 + (+element), 0);
+  const endMeetingValue = startMeetingValue + durationMeeting;
+
+  return startMeetingValue >= startDayValue && endMeetingValue <= endDayValue;
+};
+
+checksLength('string', 6);
+checksPalindrome('потоп');
+extractsDigits('333hhh333');
+checksMeeting('8:00', '17:30', '08:00', 900);
+
